requirements.txt
dj-database-url==0.5.0
Django==2.2
django-daterange-filter==1.3.0
django-environ==0.4.5
django-filter==2.1.0
django-model-utils==3.1.2
djangorestframework==3.9.2
Markdown==3.1
pytz==2019.1
sqlparse==0.3.0
